Added tests to the d3 grouped bar graph
Experimented with tape and jsdom. Not sure I'll switch to tape from jasmine, but the discovery of jsdom was interesting. Showed me how I could check actual rendering of elements on the dom in tests.

Adding tests to the kata helped, as always. Changes in this commit include changes to the application code as well, to improve testability.